requirements.txt
asgiref==3.6.0
black==23.1.0
chardet==5.1.0
click==8.1.3
distlib==0.3.6
Django==4.1.7
django-cleanup==7.0.0
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filelock==3.10.6
flake8==6.0.0
imbox==0.9.8
isort==5.12.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.2.0
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.6.0
pytz==2023.2
sqlparse==0.4.3
virtualenv==20.21.0
